plugin/xml: refactor node_map to store master in DOMImpl, not Doc
This was needed for dom_parser and doctypes to work. It does mean that dom_imple stores a global weak_ptr to the node_map from the singleton domImplementation.

Using globals makes me a bit sad, but I'm not sure there's an alternative. And the DOMImplementation is a singleton anyway, so...